branch logica para producto

+++ ButterflyBackend/src/main/java/com/proyect/Butterfly/Repositorios/VarianteRepostorio.java
@@ -3,5 +3,5 @@ package com.proyect.Butterfly.Repositorios;
 import com.proyect.Butterfly.Modelos.Variante;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VarianteRepostorio extends JpaRepository<Variante,Integer> {
+public interface VarianteRepostorio extends JpaRepository<Variante,Long> {
 }
